Corrigir tipagens do estado
